+++ fluent-react/src/index.js
@@ -20,3 +20,5 @@
 export { default as LocalizationProvider } from './provider';
 export { default as withLocalization } from './with_localization';
 export { default as Localized } from './localized';
+export { default as ReactLocalization, isReactLocalization }
+  from './localization';